SC-18&19 Moved orders from backend/backend to backend/orders
